move vmap_randomness to LossModule
